minor correction on result type

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -8,7 +8,7 @@ from typing import List
 wait_ran = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> List[int]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     '''returns a list of the outputs of wait_random'''
     res = await asyncio.gather(*(wait_ran(max_delay) for i in range(n)))
     return sorted(res)
